// src/types/api/smartContractDetailResponse.ts
/**
 * Kriptonio Private API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: private
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import type { SmartContractDeploymentResponse } from './smartContractDeploymentResponse';
import type { SmartContractTemplateResponse } from './smartContractTemplateResponse';
import type { SmartContractType } from './smartContractType';


export interface SmartContractDetailResponse { 
    id: string;
    title: string;
    compilerVersion: string;
    optimizerEnabled: boolean;
    optimizerRuns: number;
    template: SmartContractTemplateResponse;
    templateInput: any | null;
    standardJsonInput: any | null;
    type: SmartContractType;
    abi: any | null;
    bin: string;
    blockchainId: string;
    chainId: number;
    organizationId: string;
    deployment?: SmartContractDeploymentResponse;
    createdAt: string;
}
export namespace SmartContractDetailResponse {
}